+++ utilities.h
@@ -2,6 +2,6 @@
     #define UTILITIES_H
     #include <inttypes.h>
 
-    uint16_t int_pow(uint8_t a, uint8_t b);
+    uint16_t int_pow(uint16_t a, uint16_t b);
     void print_var_bits(uint8_t var);
 #endif // UTILITIES_H
